fix(sessions): satisfy boot recovery lint boundaries
